refactor: Car 생성자 정리, RacingGame 도메인 생성
