+++ src/main/java/racing/Name.java
@@ -7,7 +7,7 @@ public class Name {
 
     public Name(String name) {
         if (name.trim().length() > MAX_NAME) {
-            throw new IllegalStateException("자동차 이름은 5자를 초과할 수 없습니다.");
+            throw new IllegalStateException("자동차 이름은 " + MAX_NAME + "자를 초과할 수 없습니다.");
         }
         this.name = name;
     }
